Split vm methods by receiver type
