+++ crates/harn-vm/src/vm/methods/mod.rs
@@ -1,0 +1,10 @@
+mod dict;
+mod dispatch;
+mod generator;
+mod iter;
+mod list;
+mod number;
+mod range;
+mod set;
+mod string;
+mod struct_instance;
