+++ crates/harn-vm/src/vm/methods/number.rs
@@ -1,0 +1,12 @@
+use crate::value::{VmError, VmValue};
+
+impl crate::vm::Vm {
+    pub(super) fn call_number_method(
+        &mut self,
+        _obj: &VmValue,
+        _method: &str,
+        _args: &[VmValue],
+    ) -> Result<VmValue, VmError> {
+        Ok(VmValue::Nil)
+    }
+}
